added circle and line
git-svn-id: https://svn.princeton.edu/rowley/trunk@83 f381375e-2823-4fb2-bad6-22fa55754e8c

+++ src/RigidBody.h
@@ -4,6 +4,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
+#include <math.h>
 
 using namespace std;
 
@@ -47,7 +48,7 @@ public:
         double xc,
         double yc,
         double radius,
-        double numPoints
+        int numPoints
     );
 
     /// Add a line connecting (x1,y1) and (x2,y2)
